+++ packages/mugshot/logo.png
