+++ packages/mugshot/src/.babelrc.js
@@ -1,4 +1,4 @@
-const config = require('../.babelrc');
+const config = require('../../../.babelrc');
 
 if (process.env.COVERAGE) {
   config.plugins.push('istanbul');
