+++ packages/mugshot/src/index.ts
@@ -1,0 +1,8 @@
+import Mugshot from './lib/mugshot';
+import Browser from './interfaces/browser';
+
+export default Mugshot;
+
+export {
+  Browser
+};
