+++ packages/mugshot/src/interfaces/browser.ts
