+++ packages/mugshot/src/interfaces/file-system.ts
