+++ packages/mugshot/src/interfaces/png-differ.ts
