+++ packages/mugshot/src/lib/jimp-differ.ts
