+++ packages/webdriverio/README.md
@@ -1,0 +1,11 @@
+# `webdriverio`
+
+> TODO: description
+
+## Usage
+
+```
+const webdriverio = require('webdriverio');
+
+// TODO: DEMONSTRATE API
+```
