+++ packages/webdriverio/package.json
@@ -1,0 +1,66 @@
+{
+  "name": "@mugshot/webdriverio",
+  "publishConfig": {
+    "access": "public"
+  },
+  "version": "0.0.1",
+  "description": "A WebdriverIO adapter for Mugshot",
+  "keywords": [
+    "mugshot",
+    "webdriverio",
+    "visual",
+    "regression",
+    "testing"
+  ],
+  "author": "Andrei Picus",
+  "homepage": "https://github.com/uberVU/mugshot/tree/master/packages/webdriverio",
+  "license": "MIT",
+  "main": "dist/index",
+  "types": "dist/index",
+  "files": [
+    "dist"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/NiGhTTraX/mugshot.git"
+  },
+  "scripts": {
+    "build": "npm run clean && npm run compile",
+    "clean": "rm -rf ./dist",
+    "compile": "tsc --project ../../tsconfig.json --declaration --outDir ./dist"
+  },
+  "peerDependencies": {
+    "mugshot": "1.0.0-rc",
+    "webdriverio": "~5.8.3"
+  },
+  "devDependencies": {
+    "@babel/core": "~7.4.0",
+    "@babel/plugin-proposal-class-properties": "~7.4.0",
+    "@babel/plugin-proposal-decorators": "~7.4.0",
+    "@babel/preset-env": "~7.4.0",
+    "@babel/preset-typescript": "~7.3.0",
+    "@babel/register": "~7.4.0",
+    "@nighttrax/eslint-config-ts": "~1.0.0",
+    "@types/chai": "~4.1.7",
+    "@types/fs-extra": "~5.0.5",
+    "@types/mocha": "~5.2.5",
+    "@types/node": "~10.12.2",
+    "@typescript-eslint/eslint-plugin": "~1.5.0",
+    "babel-plugin-istanbul": "~5.1.0",
+    "chai": "~4.2.0",
+    "codecov": "~3.2.0",
+    "eslint": "~5.15.0",
+    "fs-extra": "~7.0.1",
+    "husky": "~1.3.0",
+    "jimp": "~0.6.1",
+    "lerna": "~3.13.4",
+    "lint-staged": "~8.1.0",
+    "mocha": "~6.0.2",
+    "mugshot": "1.0.0-rc",
+    "nyc": "~13.3.0",
+    "ts-node": "~8.0.0",
+    "typemoq": "~2.1.0",
+    "typescript": "~3.4.0",
+    "webdriverio": "~5.7.9"
+  }
+}
