+++ packages/webdriverio/src/.babelrc.js
@@ -1,0 +1,7 @@
+const config = require('../../../.babelrc');
+
+if (process.env.COVERAGE) {
+  config.plugins.push('istanbul');
+}
+
+module.exports = config;
