+++ packages/webdriverio/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../tsconfig.json",
+
+  "compilerOptions": {
+    "types": ["webdriver", "webdriverio"]
+  },
+
+  "include": [
+    "src/**/*"
+  ]
+}
