--- src/index.ts
@@ -1,8 +1,0 @@
-import Mugshot from './mugshot';
-import WebdriverIOAdapter from './lib/webdriverio-adapter';
-
-export default Mugshot;
-
-export {
-  WebdriverIOAdapter
-};
