+++ tests/gui/mocha.opts
@@ -1,4 +1,4 @@
 --require ts-node/register
 --recursive
 --timeout 20000
-tests/gui/**/*.spec.ts
+**/tests/gui/**/*.spec.ts
